Separated the capitalize function and tests
