+++ src/test/capitalize.test.js
@@ -1,0 +1,7 @@
+import Capital from '../modules/capitalize';
+
+const capital = Capital();
+
+test('Capitalize first letter', () => {
+  expect(capital.capitalize('martin')).toEqual('Martin');
+});
